refactor: home page refactor
