+++ apps/watch/src/components/Hero/index.ts
@@ -1,3 +1,2 @@
-export { HomeHero } from './HomeHero'
 export { SimpleHero } from './SimpleHero/SimpleHero'
 export { VideoHero } from './VideoHero/VideoHero'
