--- apps/watch/src/components/IntroText/index.ts
@@ -1,1 +1,0 @@
-export { IntroText } from './IntroText'
